// src/types/supabase.ts
export interface definitions {
  backups: {
    /**
     * Format: bigint
     * @description Note:
     * This is a Primary Key.<pk/>
     */
    id: number;
    /** Format: text */
    user_id: string;
    /** Format: integer */
    lot: number;
    /** Format: ARRAY */
    data: unknown[];
    /** Format: json */
    notes?: unknown;
  };
}